Extract batch upload composition related methods into an extension.
